osd/device: create resources unless auth list created

+++ manifests/osd/device.pp
@@ -96,6 +96,7 @@ size=1024m -n size=64k ${name}1",
 --osd-uuid ${blkid}
 ",
         creates => "${osd_data}/keyring",
+        unless  => "ceph auth list | egrep '^osd.${osd_id}$'",
         require => [
           Mount[$osd_data],
           Concat['/etc/ceph/ceph.conf'],
@@ -106,6 +107,7 @@ size=1024m -n size=64k ${name}1",
         command => "\
 ceph auth add osd.${osd_id} osd 'allow *' mon 'allow rwx' \
 -i ${osd_data}/keyring",
+        unless  => "ceph auth list | egrep '^osd.${osd_id}$'",
         require => Exec["ceph-osd-mkfs-${osd_id}"],
       }
